Movies Kata tests fixed.
